MODINV-212: Add more tests.
