+++ src/test/java/api/support/ApiTests.java
@@ -23,6 +23,7 @@ public abstract class ApiTests {
   protected final ResourceClient usersClient;
   protected final ResourceClient instancesBatchClient;
   protected final ResourceClient precedingSucceedingTitlesClient;
+  protected final ResourceClient instanceRelationshipClient;
 
   public ApiTests() {
     holdingsStorageClient = ResourceClient.forHoldingsStorage(okapiClient);
@@ -33,6 +34,7 @@ public abstract class ApiTests {
     usersClient = ResourceClient.forUsers(okapiClient);
     instancesBatchClient = ResourceClient.forInstancesBatch(okapiClient);
     precedingSucceedingTitlesClient = ResourceClient.forPrecedingSucceedingTitles(okapiClient);
+    instanceRelationshipClient = ResourceClient.forInstanceRelationship(okapiClient);
   }
 
   @BeforeClass
@@ -76,5 +78,6 @@ public abstract class ApiTests {
     preparation.deleteInstances();
 
     precedingSucceedingTitlesClient.deleteAll();
+    instanceRelationshipClient.deleteAll();
   }
 }
