+++ src/test/java/api/support/http/StorageInterfaceUrls.java
@@ -41,6 +41,10 @@ public class StorageInterfaceUrls {
     return viaOkapiURL("/preceding-succeeding-titles" + subPath);
   }
 
+  public static URL instanceRelationshipUrl(String subPath) {
+    return viaOkapiURL("/instance-storage/instance-relationships" + subPath);
+  }
+
   private static URL viaOkapiURL(String path) {
     try {
       return URLHelper.joinPath(new URL(storageOkapiUrl()), path);
